[pr] feedback from #41
* Split StreamAlert class into its own file
* Change the payload.record attribute to payload.records
